+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## Unpublished
 ### Changed
 - logger channels updated
+### Fixed
+- period - ad option overview works now with illegal(not in packages) options
 
 ## v0.6.6 - 2024-03-15
 - remove button dropdown item blade template, use dedicated css instead